Frontend - moved all pages in seperate folders for each page, move payment form to components
