+++ frontend/src/pages/UserRegisterThanks/UserRegisterThanks.css
